Modified GCompound_add to address interactor-in-compound weirdness.
If a JBECompound is added to another JBECompound, the child compound's
canvas gets set to the parent compound's canvas. This way, one can add
an interactor to a compound that isn't the top compound for the window,
and the interactor gets drawn correctly (wasn't drawn at all before,
because the interactor didn't belong to any Container). However, this
works only if the parent compound is the top compound or is a descendant
of the top compound at the time the child compound is added. I think I
also (or instead) need to add a custom setParent method to GCompound or
JBECompound.